backend/pyproject.toml
[tool.poetry]
name = "ai-recruitment-backend"
version = "0.1.0"
description = "Backend for AI Recruitment"
authors = ["Ali Mahouk"]
readme = "README.md"

[tool.poetry.dependencies]
anthropic = "^0.46.0"
azure-ai-documentintelligence = "^1.0.0"
azure-ai-vision-imageanalysis = "^1.0.0"
azure-cosmos = "^4.9.0"
azure-storage-blob = "^12.24.1"
black = "^25.1.0"
colorlog = "^6.9.0"
devtools = "^0.12.2"
fastapi = "^0.115.8"
filelock = "^3.17.0"
isort = "^6.0.0"
openai = "^1.63.2"
pillow = "^11.1.0"
pymupdf = "^1.25.3"
pypdf = "^5.3.0"
python = "^3.11"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
